- add spread

+++ app/src/main/java/au/cmcmarkets/ticker/data/api/dto/BitcoinChartPriceDTO.kt
@@ -1,9 +1,11 @@
 package au.cmcmarkets.ticker.data.api.dto
 
+import java.math.BigDecimal
+
 data class BitcoinChartPriceDTO(
-    val `15m`: Double,
-    val buy: Double,
-    val last: Double,
-    val sell: Double,
+    val `15m`: BigDecimal,
+    val buy: BigDecimal,
+    val last: BigDecimal,
+    val sell: BigDecimal,
     val symbol: String
 )
